+++ R/internal.R
@@ -1,0 +1,3 @@
+sq <- function(x) {
+  x * x
+}