+++ NEWS.md
@@ -1,3 +1,9 @@
+# uniformly 0.5.0
+
+The uniform sampling on an ellipsoid was still wrong. It is now correct, but 
+available for dimension 2 or dimension 3 only.
+
+
 # uniformly 0.3.0
 
 The uniform sampling on an ellipsoid was wrong.